fixing param setting - not sure how this was working...

+++ src/main/java/com/syarul/rnlocation/RNLocationModule.java
@@ -99,7 +99,7 @@ public class RNLocationModule extends ReactContextBaseJavaModule{
                     params.putDouble("latitude", latitude);
                     params.putDouble("speed", speed);
                     params.putDouble("altitude", altitude);
-                    params.putFloat("accuracy", accuracy);
+                    params.putDouble("accuracy", accuracy);
                     
 
                     // Send Event to JS to update Location
